// src/pages/FullAddress/FullAddress.types.ts
export type FullAddress = {
  acikAdresModel: {
    acikAdresAciklama: string;
    ada: string;
    adresNo: number;
    bagimsizBolumKayitNo: number;
    bbNitelikDetayKod: undefined;
    binaKayitNo: number;
    binaNitelikKod: number;
    binaNo: number;
    binaTur: number;
    blokAdi: string;
    csbmAdi: string;
    csbmGelismislikKod: number;
    csbmKayitNo: number;
    csbmTurKod: number;
    disKapiNo1: string;
    disKapiNo2: string;
    disKapiNoFormatted: string;
    icKapiNo: string;
    ilAdi: string;
    ilKayitNo: number;
    ilKodu: number;
    ilceAdi: string;
    ilceKayitNo: number;
    katNo: string;
    koyAdi: string;
    koyKayitNo: number;
    mahalleAdi: string;
    mahalleKayitNo: number;
    mahalleTipKod: number;
    maksBbDurum: number;
    maksBinaDurum: number;
    maksBinaNumaratajTipi: number;
    maksBinaYapiTipi: number;
    mernisIlceKod: number;
    pafta: string;
    parsel: string;
    postaKodu: undefined;
    siteAdi: undefined;
    yapiKullanimAmac: number;
    yapiKullanimAmacBsKod: number;
  };
  ada: string;
  adi: string;
  adresNo: number;
  bilesenAdi: string;
  binaKayitNo: number;
  binaNo: number;
  blokAdi: string;
  disKapiNo: string;
  icKapiNo: string;
  katNo: string;
  kimlikNo: number;
  maksBbDurum: number;
  maksBbDurumFormatted: string;
  maksBbTip: number;
  maksBbTipFormatted: string;
  maksBinaNumaratajTipi: number;
  maksBinaNumaratajTipiFormatted: string;
  pafta: string;
  parsel: string;
  postaKodu: undefined;
  siteAdi: undefined;
  tapuBagimsizBolumNo: undefined;
  yapiKullanimAmac: number;
  yapiKullanimAmacFormatted: string;
};